+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 - [core] support #if(){} syntax.
 - [core] switch a large amount of class modifier from public to internal.
 - [core] optimized performance for nested sql.
+- [vstool] improve performance.
 
 ## 0.9.1:
 - [core] fix a issue that EnsureCompiled will fail when using sub-sql.